requirements.txt





aniso8601==10.0.0
bcrypt==4.2.1
blinker==1.9.0
celery[Redis]==5.3.4
celery==5.3.4
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email-validator==2.1.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask==3.1.0
greenlet==3.1.1
gunicorn==21.2.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pandas==2.1.1
passlib==1.7.4
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2025.1
redis==4.5.4
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug>=3.1
WTForms==3.2.1